+++ src/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from "vscode";
 
 import { checkForExerciseUpdates, checkForNewExercises } from "./actions";
 import TMC from "./api/tmc";
+import VSC, { showError } from "./api/vscode";
 import WorkspaceManager from "./api/workspaceManager";
 import { EXERCISE_CHECK_INTERVAL } from "./config/constants";
 import Settings from "./config/settings";
@@ -11,7 +12,6 @@ import { validateAndFix } from "./config/validate";
 import * as init from "./init";
 import TemporaryWebviewProvider from "./ui/temporaryWebviewProvider";
 import UI from "./ui/ui";
-import { showError } from "./utils/";
 import Logger from "./utils/logger";
 
 let maintenanceInterval: NodeJS.Timeout | undefined;
@@ -19,7 +19,9 @@ let maintenanceInterval: NodeJS.Timeout | undefined;
 export async function activate(context: vscode.ExtensionContext): Promise<void> {
     const logger = new Logger();
     logger.log(`Starting extension in "${process.env.DEBUG_MODE || "production"}" mode.`);
+
     const storage = new Storage(context);
+
     const resourcesResult = await init.resourceInitialization(context, storage, logger);
     if (resourcesResult.err) {
         const message = `TestMyCode Initialization failed: ${resourcesResult.val}`;
@@ -32,19 +34,17 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
     const settingsResult = await init.settingsInitialization(storage, resources, logger);
     const settings = new Settings(storage, logger, settingsResult, resources);
     logger.setLogLevel(settings.getLogLevel());
-    await vscode.commands.executeCommand("setContext", "tmcWorkspaceActive", true);
 
-    logger.log(`VSCode version: ${vscode.version}`);
+    const vsc = new VSC(settings, logger);
+    await vsc.activate();
+
+    logger.log(`VSCode version: ${vsc.getVSCodeVersion()}`);
     logger.log(`TMC extension version: ${resources.extensionVersion}`);
-    logger.log(
-        `Python extension version: ${
-            vscode.extensions.getExtension("ms-python.python")?.packageJSON.version
-        }`,
-    );
+    logger.log(`Python extension version: ${vsc.getExtensionVersion("ms-python.python")}`);
 
     const ui = new UI(context, resources, vscode.window.createStatusBarItem());
-
     const tmc = new TMC(storage, resources, logger);
+
     const validationResult = await validateAndFix(storage, tmc, ui, resources, logger);
     if (validationResult.err) {
         const message = `Data reconstruction failed: ${validationResult.val.message}`;
@@ -63,6 +63,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
         settings,
         temporaryWebviewProvider,
         tmc,
+        vsc,
         ui,
         userData,
         workspaceManager,