+++ src/init/ui.ts
@@ -17,15 +17,9 @@ import {
     updateCourse,
 } from "../actions";
 import { ActionContext, CourseExerciseDownloads } from "../actions/types";
+import { askForConfirmation, showError, showNotification } from "../api/vscode";
 import { ExerciseStatus } from "../config/types";
-import {
-    askForConfirmation,
-    isWorkspaceOpen,
-    LogLevel,
-    showError,
-    showNotification,
-    sleep,
-} from "../utils/";
+import { isWorkspaceOpen, LogLevel, sleep } from "../utils/";
 
 /**
  * Registers the various actions and handlers required for the user interface to function.
@@ -154,20 +148,26 @@ export function registerUiActions(actionContext: ActionContext): void {
             }
         },
     );
-    ui.webview.registerHandler("courseDetails", (msg: { type?: "courseDetails"; id?: number }) => {
-        if (!(msg.type && msg.id)) {
-            return;
-        }
-        const courseId: number = msg.id;
-        displayLocalCourseDetails(actionContext, msg.id);
-        const uiState = ui.webview.getStateId();
-        // Try to fetch updates from API
-        updateCourse(actionContext, courseId).then(() =>
-            uiState === ui.webview.getStateId()
-                ? displayLocalCourseDetails(actionContext, courseId)
-                : {},
-        );
-    });
+    ui.webview.registerHandler(
+        "courseDetails",
+        async (msg: { type?: "courseDetails"; id?: number; useCache?: boolean }) => {
+            if (!(msg.type && msg.id)) {
+                return;
+            }
+            const courseId: number = msg.id;
+            const uiState = ui.webview.getStateId();
+
+            if (msg.useCache) {
+                displayLocalCourseDetails(actionContext, courseId);
+            } else {
+                updateCourse(actionContext, courseId).then(() =>
+                    uiState === ui.webview.getStateId()
+                        ? displayLocalCourseDetails(actionContext, courseId)
+                        : {},
+                );
+            }
+        },
+    );
     ui.webview.registerHandler(
         "openSelected",
         async (msg: { type?: "openSelected"; ids?: number[] }) => {