+++ src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from "./dateDeadline";
-export * from "./dialog";
 export * from "./env";
 export * from "./utils";
 export * from "./logger";